+++ src/lib/string/memcpy.c
@@ -11,11 +11,11 @@
 #include <string.h>
 
 /* How many bytes are copied each iteration of the word copy loop.  */
-#define BLOCK_SZ (sizeof(long))
+#define BLOCK_SZ (sizeof(int))
 
 /* Nonzero if either X or Y is not aligned on a "long" boundary.  */
 #define unaligned(x, y) \
-  (((long) x | (long) y) & (sizeof(long) - 1))
+  (((int) x | (int) y) & (sizeof(int) - 1))
 
 void *memcpy(void *dst_, const void *src_, size_t n) {
 	char *dst = dst_;
